deepseek-ocr/requirements.txt
accelerate==1.11.0
addict==2.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
deprecated==1.2.18
easydict==1.13
einops==0.8.1
fastapi==0.120.0
filelock==3.19.1
flash-attn==2.7.3
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
img2pdf==0.6.1
jinja2==3.1.6
lxml==6.0.2
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==25.0
pikepdf==10.0.0
pillow==11.3.0
psutil==7.1.2
pydantic==2.12.3
pydantic-core==2.41.4
pymupdf==1.26.5
python-multipart==0.0.20
pyyaml==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
wheel==0.45.1
wrapt==1.17.3
